Refactors demo files from master changes
